[web] adding a test for e2e web testing. (#2554)
* adding a test for e2e web testing.

* chaning the changelog. updating pubspec.yaml version. fixing analyze

* merging the changelog

* addressing reviewer comments

* fix format. addressing reviewer comments

* try to run chromedriver on the backend

* chrome driver is was running as the main task, preventing test from running . use background_task to run it. use ENV for directory change

* change in scripts. remove list from backfround task. change argument of driver

* removed background task. it wasn't using the same path

* run web tests only on browser

* add test on to the web driver test as well. drive-examples are still failing

* fix the imports

* trying reviever suggestion for drive_example compile errors

* test _ imports

* Revert "test _ imports"

This reverts commit 7cf24d6cb5833efa437a78cf5b843eb6b6368a81.

* removing the web_test driver file upon reviewers suggestion

* format fix
