+++ packages/e2e/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.4+4
+
+* Fixed a hang that occurred on platforms that don't have a `MethodChannel` listener registered..
+
 ## 0.2.4+3
 
 * Fixed code snippet in the readme under the "Using Flutter driver to run tests" section.
