+++ packages/e2e/example/test_driver/failure.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:e2e/e2e.dart';
 
-import 'package:e2e_example/main.dart';
+import 'package:e2e_example/main.dart' as app;
 
 // Tests the failure behavior of the E2EWidgetsFlutterBinding
 //
@@ -21,10 +21,10 @@ void main() {
 
   testWidgets('failure 1', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(MyApp());
+    app.main();
 
     // Verify that platform version is retrieved.
-    expect(
+    await expectLater(
       find.byWidgetPredicate(
         (Widget widget) =>
             widget is Text && widget.data.startsWith('This should fail'),
